tweak how benchmarks are run
